Fix warnings in Xcode.

+++ cstructs/memprofile.h
@@ -10,8 +10,8 @@ void printmeminfo();
 
 #if 1
 
-#define malloc(numBytes) memop(__FILE__, __LINE__, NULL, numBytes, 0)
-#define realloc(oldPtr, numBytes) memop(__FILE__, __LINE__, oldPtr, numBytes, 1)
+#define malloc(numBytes) memop(__FILE__, __LINE__, NULL, (int)numBytes, 0)
+#define realloc(oldPtr, numBytes) memop(__FILE__, __LINE__, oldPtr, (int)numBytes, 1)
 #define free(ptr) memop(__FILE__, __LINE__, ptr, -1, 0)
 
 #endif
